fix: harden dashboard project selection
